refactor: Question 엔티티의 원시값, 컬렉션 포장
